% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage data set}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{the distance from the fire station, miles}
  \item{DAMAGE}{the damage caused, $'000}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set of the distance from a fire and the damage
caused at those locations.
}
\keyword{datasets}
