% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinomial.r
\name{mybinomial}
\alias{mybinomial}
\title{Binomial distribution simulator}
\usage{
mybinomial(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{The number of iterations to simulate, an integer}

\item{n}{The size of the binomial distribution}

\item{p}{The probability of success}
}
\value{
A barplot of the binomial distribution
}
\description{
Binomial distribution simulator
}
\examples{
\dontrun{mybin(iter = 10000, n = 20, p = 0.7)}
}
