% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{My CI bootstrap function
Bootstraps a (100%-alpha) confidence interval and
displays a histogram with confidence intervals and the mean.}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The number of iterations to run.}

\item{x}{A numeric vector of the sample.}

\item{fun}{The function to produce the statistic to be tested.}

\item{alpha}{1-(percent confidence interval)}

\item{cx}{Numeric character expansion factor to be passed to text()}

\item{...}{Additional parameters passed to hist()}
}
\value{
A list consisting of confidence intervals, the sample, the statistic, and the provided function
}
\description{
My CI bootstrap function
Bootstraps a (100%-alpha) confidence interval and
displays a histogram with confidence intervals and the mean.
}
\examples{
\dontrun{
sam <- rnorm(20, mean=4, sd=2)
myboot(iter=10000, x=sam, fun="mean", alpha=0.05)
}
}
