% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Central Limit Function, Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{the number of samples that should be taken each iteration}

\item{iter}{the number of iterations}

\item{lambda}{the lambda value of the poisson distribution, defaults to 10}

\item{...}{additional options to be passed to the histogram function}
}
\value{
a vector of all of the samples taken
}
\description{
My Central Limit Function, Poisson Distribution
}
\examples{
\dontrun{mycltp(n=10, iter=10000, lambda=5)}
}
