% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{"My DDT" function}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{- the data frame. This should be ddt.}

\item{SPECIES}{- the species of fish to filter for.}
}
\value{
A LENGTH vs WEIGHT plot and a named list with the pre- and post-filtering data frame as well as a frequency table of rivers.
}
\description{
Creates a ggplot of the fish SPECIES, length vs weight, coloured by river.
Saves the data, filtered to a specific species, to the working directory.
Creates a list with three attributes:
}
\details{
\describe{
  \item{pre_subset}{the data frame passed to the function (the `df` parameter).}
  \item{post_subset}{the data frame filtered by `SPECIES`.}
  \item{river_freq}{a frequency table of `RIVER`s for the given `SPECIES`.}
}
}
\examples{
\dontrun{myddt(df=ddt, SPECIES="CCATFISH")}
}
