% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{My Maximum Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{- a function that will find the log likelihood given a value-parameter pair}

\item{x}{- a vector of prior outcomes}

\item{param}{- the parameter to be passed to the distribution (for instance, p in a binomial or lambda in a Poisson)}

\item{...}{- other parameters that will be passed to `plot()`}
}
\value{
maxima and also a neat graph
}
\description{
Uses the method of maximum likelihood to estimate the max likelihood
of a distribution, given some prior outcomes and a function that
returns the log likelihood given a parameter.
}
\examples{
\dontrun{
mymaxlik(function (x, param) log(dbinom(x, prob=param, size=20)), c(1,2,3,4), seq(0,1,length=1000))
}
}
