% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal curve plotter and probability finder}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The number we will find the probability P(x <= a)}
}
\value{
The probability P(x <= a)
}
\description{
Normal curve plotter and probability finder
}
\examples{
\dontrun{myncurve(mu=0, sigma=1, a=0)}
}
