% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.r
\name{scatterhist}
\alias{scatterhist}
\title{Scatter Histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{the x parameter, a vector}

\item{y}{the y parameter, a vector}

\item{xlab}{the x label}

\item{ylab}{the y label}
}
\value{
A plot, a scatter histogram
}
\description{
Scatter Histogram
}
\examples{
\dontrun{with(ddt, scatterhist(LENGTH,WEIGHT, xlab="LENGTH"))}
}
